refactor(mcp): Extract tools into individual files
