+++ web/src/features/mcp/features/annotationQueues/index.ts
@@ -1,26 +1,44 @@
 import type { McpFeatureModule } from "../../server/registry";
 import {
   createAnnotationQueueAssignmentTool,
-  createAnnotationQueueItemTool,
-  createAnnotationQueueTool,
-  deleteAnnotationQueueAssignmentTool,
-  deleteAnnotationQueueItemTool,
-  getAnnotationQueueItemTool,
-  getAnnotationQueueTool,
-  handleCreateAnnotationQueue,
   handleCreateAnnotationQueueAssignment,
+} from "./tools/createAnnotationQueueAssignment";
+import {
+  createAnnotationQueueItemTool,
   handleCreateAnnotationQueueItem,
+} from "./tools/createAnnotationQueueItem";
+import {
+  createAnnotationQueueTool,
+  handleCreateAnnotationQueue,
+} from "./tools/createAnnotationQueue";
+import {
+  deleteAnnotationQueueAssignmentTool,
   handleDeleteAnnotationQueueAssignment,
+} from "./tools/deleteAnnotationQueueAssignment";
+import {
+  deleteAnnotationQueueItemTool,
   handleDeleteAnnotationQueueItem,
-  handleGetAnnotationQueue,
+} from "./tools/deleteAnnotationQueueItem";
+import {
+  getAnnotationQueueItemTool,
   handleGetAnnotationQueueItem,
+} from "./tools/getAnnotationQueueItem";
+import {
+  getAnnotationQueueTool,
+  handleGetAnnotationQueue,
+} from "./tools/getAnnotationQueue";
+import {
   handleListAnnotationQueueItems,
-  handleListAnnotationQueues,
-  handleUpdateAnnotationQueueItem,
   listAnnotationQueueItemsTool,
+} from "./tools/listAnnotationQueueItems";
+import {
+  handleListAnnotationQueues,
   listAnnotationQueuesTool,
+} from "./tools/listAnnotationQueues";
+import {
+  handleUpdateAnnotationQueueItem,
   updateAnnotationQueueItemTool,
-} from "./tools";
+} from "./tools/updateAnnotationQueueItem";
 
 export const annotationQueuesFeature: McpFeatureModule = {
   name: "annotationQueues",
