+++ web/src/features/mcp/features/comments/tools/index.ts
@@ -1,0 +1,3 @@
+export * from "./createComment";
+export * from "./getComment";
+export * from "./listComments";
