+++ web/src/features/mcp/features/datasets/tools/index.ts
@@ -1,0 +1,12 @@
+export * from "./createDataset";
+export * from "./createDatasetItem";
+export * from "./createDatasetRunItem";
+export * from "./deleteDatasetItem";
+export * from "./deleteDatasetRun";
+export * from "./getDataset";
+export * from "./getDatasetItem";
+export * from "./getDatasetRun";
+export * from "./listDatasetItems";
+export * from "./listDatasetRunItems";
+export * from "./listDatasetRuns";
+export * from "./listDatasets";
