+++ web/src/features/mcp/features/health/tools/index.ts
@@ -1,0 +1,1 @@
+export * from "./getHealth";
