+++ web/src/features/mcp/features/models/tools/createModel.ts
@@ -1,93 +1,16 @@
-import { InvalidRequestError, LangfuseNotFoundError } from "@langfuse/shared";
+import { InvalidRequestError } from "@langfuse/shared";
 import { Prisma, prisma } from "@langfuse/shared/src/db";
 import { clearModelCacheForProject } from "@langfuse/shared/src/server";
 import { auditLog } from "@/src/features/audit-logs/auditLog";
 import { isValidPostgresRegex } from "@/src/features/models/server/isValidPostgresRegex";
 import {
-  DeleteModelV1Query,
-  DeleteModelV1Response,
-  GetModelV1Query,
-  GetModelV1Response,
-  GetModelsV1Query,
-  GetModelsV1Response,
   PostModelsV1Body,
   PostModelsV1Response,
   prismaToApiModelDefinition,
 } from "@/src/features/public-api/types/models";
-import { defineTool } from "../../core/define-tool";
-import { runMcpTool } from "../../core/run-mcp-tool";
-import { paginationMeta } from "../publicApi";
-
-const modelPricingInclude = {
-  pricingTiers: {
-    select: {
-      id: true,
-      name: true,
-      isDefault: true,
-      priority: true,
-      conditions: true,
-      prices: {
-        select: {
-          usageType: true,
-          price: true,
-        },
-      },
-    },
-    orderBy: { priority: "asc" as const },
-  },
-};
-
-export const [listModelsTool, handleListModels] = defineTool({
-  name: "listModels",
-  description:
-    "List custom and Langfuse-managed model definitions visible to the current project.",
-  baseSchema: GetModelsV1Query,
-  inputSchema: GetModelsV1Query,
-  handler: async (input, context) =>
-    runMcpTool({
-      spanName: "mcp.models.list",
-      context,
-      attributes: {
-        "mcp.pagination_page": input.page,
-        "mcp.pagination_limit": input.limit,
-      },
-      fn: async () => {
-        const where = {
-          OR: [{ projectId: context.projectId }, { projectId: null }],
-        };
-
-        const [models, totalItems] = await Promise.all([
-          prisma.model.findMany({
-            where,
-            orderBy: [
-              { modelName: "asc" },
-              { unit: "asc" },
-              {
-                startDate: {
-                  sort: "desc",
-                  nulls: "last",
-                },
-              },
-            ],
-            include: modelPricingInclude,
-            take: input.limit,
-            skip: (input.page - 1) * input.limit,
-          }),
-          prisma.model.count({ where }),
-        ]);
-
-        return GetModelsV1Response.parse({
-          data: models.map(prismaToApiModelDefinition),
-          meta: paginationMeta({
-            page: input.page,
-            limit: input.limit,
-            totalItems,
-          }),
-        });
-      },
-    }),
-  readOnlyHint: true,
-});
+import { defineTool } from "../../../core/define-tool";
+import { runMcpTool } from "../../../core/run-mcp-tool";
+import { modelPricingInclude } from "../schema";
 
 export const [createModelTool, handleCreateModel] = defineTool({
   name: "createModel",
@@ -226,89 +149,4 @@ export const [createModelTool, handleCreateModel] = defineTool({
         );
       },
     }),
-});
-
-export const [getModelTool, handleGetModel] = defineTool({
-  name: "getModel",
-  description: "Get a model definition by ID from the current project scope.",
-  baseSchema: GetModelV1Query,
-  inputSchema: GetModelV1Query,
-  handler: async (input, context) =>
-    runMcpTool({
-      spanName: "mcp.models.get",
-      context,
-      attributes: { "mcp.model_id": input.modelId },
-      fn: async () => {
-        const model = await prisma.model.findFirst({
-          where: {
-            AND: [
-              { id: input.modelId },
-              {
-                OR: [{ projectId: context.projectId }, { projectId: null }],
-              },
-            ],
-          },
-          include: modelPricingInclude,
-        });
-
-        if (!model) {
-          throw new LangfuseNotFoundError("No model with this id found.");
-        }
-
-        return GetModelV1Response.parse(prismaToApiModelDefinition(model));
-      },
-    }),
-  readOnlyHint: true,
-});
-
-export const [deleteModelTool, handleDeleteModel] = defineTool({
-  name: "deleteModel",
-  description:
-    "Delete a custom model definition from the current project. Built-in models cannot be deleted.",
-  baseSchema: DeleteModelV1Query,
-  inputSchema: DeleteModelV1Query,
-  handler: async (input, context) =>
-    runMcpTool({
-      spanName: "mcp.models.delete",
-      context,
-      attributes: { "mcp.model_id": input.modelId },
-      fn: async () => {
-        const model = await prisma.model.findFirst({
-          where: {
-            id: input.modelId,
-            projectId: context.projectId,
-          },
-        });
-
-        if (!model) {
-          throw new LangfuseNotFoundError(
-            "No model with this id found. Note: You cannot delete built-in models, override them with a model with the same name.",
-          );
-        }
-
-        await prisma.model.delete({
-          where: {
-            id: input.modelId,
-            projectId: context.projectId,
-          },
-        });
-
-        await auditLog({
-          action: "delete",
-          resourceType: "model",
-          resourceId: input.modelId,
-          projectId: context.projectId,
-          orgId: context.orgId,
-          apiKeyId: context.apiKeyId,
-          before: model,
-        });
-
-        await clearModelCacheForProject(context.projectId);
-
-        return DeleteModelV1Response.parse({
-          message: "Model successfully deleted",
-        });
-      },
-    }),
-  destructiveHint: true,
 });
